requirements.txt
passlib[bcrypt]
fastapi[standard]
sqlmodel
psycopg2
pyjwt
python-jose
pytz
uvicorn